Create blueprints
This allows us to add more routes to our app without bloating the
app/__init__.py file. We can separate like routes to the same file and
keep the app cleaner.
