+++ artistic/views/__init__.py
@@ -1,0 +1,2 @@
+from artistic.views.auth import auth_bp
+from artistic.views.home import home_bp
